pyproject.toml
[project]
name = "libraryms"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.1",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "djoser>=2.3.1",
    "drf-nested-routers>=0.94.2",
    "drf-yasg>=1.21.10",
    "mysqlclient>=2.2.7",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.0",
    "routers>=0.10.1",
]
